+++ stackslib/src/net/api/tests/blockreplay.rs
@@ -19,7 +19,7 @@ use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use stacks_common::types::chainstate::StacksBlockId;
 
 use crate::chainstate::stacks::Error as ChainError;
-use crate::net::api::blocksimulate;
+use crate::net::api::blockreplay;
 use crate::net::api::tests::TestRPC;
 use crate::net::connection::ConnectionOptions;
 use crate::net::httpcore::{StacksHttp, StacksHttpRequest};
@@ -32,7 +32,7 @@ fn test_try_parse_request() {
     let mut http = StacksHttp::new(addr.clone(), &ConnectionOptions::default());
 
     let mut request =
-        StacksHttpRequest::new_block_simulate(addr.into(), &StacksBlockId([0x01; 32]));
+        StacksHttpRequest::new_block_replay(addr.into(), &StacksBlockId([0x01; 32]));
 
     // add the authorization header
     request.add_header("authorization".into(), "password".into());
@@ -44,7 +44,7 @@ fn test_try_parse_request() {
     let (parsed_preamble, offset) = http.read_preamble(&bytes).unwrap();
 
     let mut handler =
-        blocksimulate::RPCNakamotoBlockSimulateRequestHandler::new(Some("password".into()));
+        blockreplay::RPCNakamotoBlockReplayRequestHandler::new(Some("password".into()));
 
     let mut parsed_request = http
         .handle_try_parse_request(
@@ -67,7 +67,7 @@ fn test_try_parse_request() {
 #[test]
 fn test_block_reply_errors() {
     let mut handler =
-        blocksimulate::RPCNakamotoBlockSimulateRequestHandler::new(Some("password".into()));
+        blockreplay::RPCNakamotoBlockReplayRequestHandler::new(Some("password".into()));
 
     let test_observer = TestEventObserver::new();
     let mut rpc_test = TestRPC::setup_nakamoto(function_name!(), &test_observer);
@@ -101,21 +101,21 @@ fn test_try_make_response() {
 
     // query existing, non-empty Nakamoto block
     let mut request =
-        StacksHttpRequest::new_block_simulate(addr.clone().into(), &rpc_test.canonical_tip);
+        StacksHttpRequest::new_block_replay(addr.clone().into(), &rpc_test.canonical_tip);
     // add the authorization header
     request.add_header("authorization".into(), "password".into());
     requests.push(request);
 
     // query non-existent block
     let mut request =
-        StacksHttpRequest::new_block_simulate(addr.clone().into(), &StacksBlockId([0x01; 32]));
+        StacksHttpRequest::new_block_replay(addr.clone().into(), &StacksBlockId([0x01; 32]));
     // add the authorization header
     request.add_header("authorization".into(), "password".into());
     requests.push(request);
 
     // unauthenticated request
     let request =
-        StacksHttpRequest::new_block_simulate(addr.clone().into(), &StacksBlockId([0x00; 32]));
+        StacksHttpRequest::new_block_replay(addr.clone().into(), &StacksBlockId([0x00; 32]));
     requests.push(request);
 
     let mut responses = rpc_test.run(requests);
@@ -128,7 +128,7 @@ fn test_try_make_response() {
         std::str::from_utf8(&response.try_serialize().unwrap()).unwrap()
     );
 
-    let resp = response.decode_simulated_block().unwrap();
+    let resp = response.decode_replayed_block().unwrap();
 
     let tip_block = test_observer.get_blocks().last().unwrap().clone();