+++ stackslib/src/net/api/tests/mod.rs
@@ -59,7 +59,7 @@ use crate::net::{
     UrlString,
 };
 
-mod blocksimulate;
+mod blockreplay;
 mod callreadonly;
 mod fastcallreadonly;
 mod get_tenures_fork_info;